Extract 'MediaStoreSpy' into a shared scope to remove duplication
