+++ MovieListFrameworkTests/Media Cache/Helpers/MediaStoreSpy.swift
@@ -1,0 +1,47 @@
+//
+//  MediaStoreSpy.swift
+//  MovieListFrameworkTests
+//
+//  Created by macbook on 13/09/2023.
+//
+
+import Foundation
+import MovieListFramework
+
+class MediaStoreSpy: MediaStore {
+    enum ReceivedMessage: Equatable {
+        case deleteCachedFeed
+        case insert([LocalMediaItem], Date)
+    }
+    
+    private(set) var receivedMessages = [ReceivedMessage]()
+    
+    private var deletionCompletions = [DeletionCompletion]()
+    private var insertionCompletions = [InsertionCompletion]()
+    
+    func deleteCachedMedia(completion: @escaping DeletionCompletion) {
+        deletionCompletions.append(completion)
+        receivedMessages.append(.deleteCachedFeed)
+    }
+    
+    func completeDeletion(with error: Error, at index: Int = 0) {
+        deletionCompletions[index](error)
+    }
+    
+    func completeDeletionSuccessfully(at index: Int = 0) {
+        deletionCompletions[index](nil)
+    }
+    
+    func insert(_ items: [LocalMediaItem], timestamp: Date, completion: @escaping (InsertionCompletion)) {
+        insertionCompletions.append(completion)
+        receivedMessages.append(.insert(items, timestamp))
+    }
+    
+    func completeInsertion(with error: Error, at index: Int = 0) {
+        insertionCompletions[index](error)
+    }
+    
+    func completeInsertionSuccessfully(at index: Int = 0) {
+        insertionCompletions[index](nil)
+    }
+}
